requirements.txt
altair==5.5.0
docxtpl==0.19.1
easyocr==1.7.2
matplotlib==3.10.1
numpy==2.2.3
opencv_python_headless==4.11.0.86
pandas==2.2.3
Pillow==11.1.0
scikit_learn==1.6.1
streamlit==1.39.0
torch==2.6.0+cu126
torchvision==0.21.0
